services/flask-backend/requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
werkzeug==3.0.2
